refactor(tui): split test modules out of two thousand-line files
native_memory.rs (1091 lines) and tui/views/fleet_roster.rs (1044 lines)
crossed the 1000-line production-module ceiling in
scripts/source-structure-budget.json during the v0.9.4 train.

Move each file's #[cfg(test)] module verbatim into a Rust-2018 sibling
submodule (native_memory/tests.rs, fleet_roster/tests.rs); test files are
excluded from the budget's production-source accounting by definition. The
only non-verbatim change is the include! path in the fleet roster tests,
which now resolves one directory up. Public APIs are untouched; both files
are back under the ceiling (773 and 638 lines).
